requirements.txt
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.24.0
fredapi>=0.5.1
anthropic>=0.39.0
jinja2>=3.1.2
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.4
